requirements.txt
click
httpx
asyncpg
python-dotenv
pydantic
redis[async]
ebooklib
beautifulsoup4
Pillow
pyyaml
easyocr
numpy
torch
